Created new file structure using agreed upon filestructure from discord + some new changes like /core/ containing facade. added facade layer Removed need for productDTO from project, instead using productresponse. added a product mapper to make controller extra lean :)
