+++ src/main/kotlin/org/ptss/support/api/dtos/requests/CreateProductRequest.kt
@@ -1,0 +1,7 @@
+package org.ptss.support.api.dtos.requests
+
+data class CreateProductRequest(
+    val name: String,
+    val description: String,
+    val media: List<String>
+)
