+++ src/main/kotlin/org/ptss/support/api/dtos/responses/ProductResponse.kt
@@ -1,0 +1,8 @@
+package org.ptss.support.api.dtos.responses
+
+data class ProductResponse(
+    val id: String,
+    val name: String,
+    val description: String,
+    val media: List<String>
+)
