--- src/main/kotlin/org/ptss/support/cqrs/core/commands/product/CreateProductCommand.kt
@@ -1,8 +1,0 @@
-package org.ptss.support.cqrs.core.commands.product
-
-data class CreateProductCommand(
-    val name: String,
-    val description: String,
-    val media: List<String>
-)
-
