--- src/main/kotlin/org/ptss/support/cqrs/core/dtos/ProductDto.kt
@@ -1,8 +1,0 @@
-package org.ptss.support.cqrs.core.dtos
-
-data class ProductDto(
-    val name: String,
-    val description: String,
-    val media: List<String>
-)
-
