--- src/main/kotlin/org/ptss/support/cqrs/core/interfaces/ICommandHandler.kt
@@ -1,5 +1,0 @@
-package org.ptss.support.cqrs.core.interfaces
-
-interface ICommandHandler<in TCommand, TResult> {
-    suspend fun handleAsync(command: TCommand): TResult
-}
