--- src/main/kotlin/org/ptss/support/cqrs/core/interfaces/IQueryHandler.kt
@@ -1,5 +1,0 @@
-package org.ptss.support.cqrs.core.interfaces
-
-interface IQueryHandler<in TQuery, TResult> {
-    suspend fun handleAsync(query: TQuery): TResult
-}
