--- src/main/kotlin/org/ptss/support/cqrs/core/models/Product.kt
@@ -1,10 +1,0 @@
-package org.ptss.support.cqrs.core.models
-
-import java.util.UUID
-
-data class Product(
-    val id: String = UUID.randomUUID().toString(),
-    val name: String,
-    val description: String,
-    val media: List<String> = emptyList()
-)
