--- src/main/kotlin/org/ptss/support/cqrs/core/queries/product/GetAllProductsQuery.kt
@@ -1,3 +1,0 @@
-package org.ptss.support.cqrs.core.queries.product
-
-class GetAllProductsQuery()
