--- src/main/kotlin/org/ptss/support/cqrs/core/queries/product/GetProductByIdQuery.kt
@@ -1,3 +1,0 @@
-package org.ptss.support.cqrs.core.queries.product
-
-data class GetProductByIdQuery(val id: String)
