--- src/main/kotlin/org/ptss/support/cqrs/infrastructure/interfaces/IProductRepository.kt
@@ -1,9 +1,0 @@
-package org.ptss.support.cqrs.infrastructure.interfaces
-
-import org.ptss.support.cqrs.core.models.Product
-
-interface IProductRepository {
-    fun create(product: Product): String
-    fun getById(id: String): Product?
-    fun getAll(): List<Product>
-}
