--- src/main/kotlin/org/ptss/support/cqrs/infrastructure/util/ExceptionLogging.kt
@@ -1,17 +1,0 @@
-package org.ptss.support.cqrs.infrastructure.util
-
-import org.slf4j.Logger
-
-suspend fun <T> Logger.executeWithExceptionLoggingAsync(
-    operation: suspend () -> T,
-    logMessage: String,
-    exceptionHandling: ((Exception) -> Exception)? = null,
-    vararg args: Any
-): T {
-    return try {
-        operation()
-    } catch (ex: Exception) {
-        this.error(logMessage.format(*args), ex)
-        throw exceptionHandling?.invoke(ex) ?: ex
-    }
-}
