+++ src/main/kotlin/org/ptss/support/domain/commands/CreateProductCommand.kt
@@ -1,0 +1,8 @@
+package org.ptss.support.domain.commands
+
+data class CreateProductCommand(
+    val name: String,
+    val description: String,
+    val media: List<String>
+)
+
