+++ src/main/kotlin/org/ptss/support/domain/interfaces/cqrs/ICommandHandler.kt
@@ -1,0 +1,5 @@
+package org.ptss.support.domain.interfaces.cqrs
+
+interface ICommandHandler<in TCommand, TResult> {
+    suspend fun handleAsync(command: TCommand): TResult
+}
