+++ src/main/kotlin/org/ptss/support/domain/interfaces/cqrs/IQueryHandler.kt
@@ -1,0 +1,5 @@
+package org.ptss.support.domain.interfaces.cqrs
+
+interface IQueryHandler<in TQuery, TResult> {
+    suspend fun handleAsync(query: TQuery): TResult
+}
