+++ src/main/kotlin/org/ptss/support/domain/querries/GetAllProductsQuery.kt
@@ -1,0 +1,3 @@
+package org.ptss.support.domain.querries
+
+class GetAllProductsQuery()
