+++ src/main/kotlin/org/ptss/support/domain/querries/GetProductByIdQuery.kt
@@ -1,0 +1,3 @@
+package org.ptss.support.domain.querries
+
+data class GetProductByIdQuery(val id: String)
