--- src/main/kotlin/org/ptss/support/exceptions/APIException.kt
@@ -1,10 +1,0 @@
-package org.ptss.support.exceptions
-
-import jakarta.xml.bind.ValidationException
-import org.ptss.support.enums.ErrorCode
-
-class APIException(
-    override val message: String,
-    val errorCode: ErrorCode,
-    val details: ErrorDetails? = null
-) : ValidationException(message)
